stdio: Fix -Wtautological-constant-out-of-range-compare on clang

+++ stdio-common/vfprintf-internal.c
@@ -1878,7 +1878,7 @@ printf_positional (FILE *s, const CHAR_T *format, int readonly_format,
 	  extern printf_function **__printf_function_table;
 	  int function_done;
 
-	  if (spec <= UCHAR_MAX
+	  if ((int) spec <= UCHAR_MAX
 	      && __printf_function_table != NULL
 	      && __printf_function_table[(size_t) spec] != NULL)
 	    {
